+++ RELEASES.md
@@ -6,6 +6,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template: tmpl:ID attribute, and capturing unused vars
 * prov-template: fixed isue with groups
 * prov-template: preparing to replace oldBindings by new Bindings
+* prov-template: WIP: running new Bindings in parallel with OldBindings, and comparing
 
 `2.2.2`
 * prov-template: added ptm-merging template